update: module, pydub
